requirements.txt
pytest==7.3.1
pytest-html==3.2.0
pytest-metadata==2.0.4
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==1.3.1
urllib3==2.2.1
WooCommerce==3.0.0
selenium~=4.23.1
allure-pytest
configobj~=5.0.8
PyMySQL~=1.1.1
setuptools~=69.2.0